+++ pages/boards.tsx
@@ -1,7 +1,6 @@
 import axios from "@/lib/axios";
 import Image from "next/image";
 import formatDate from "@/util/formatDate";
-import Badge from "@/assets/images/ui/best_badge.svg";
 import Heart from "@/assets/images/icons/ic_heart.svg";
 import Search from "@/assets/images/icons/ic_search.svg";
 import Profile from "@/assets/images/icons/ic_profile.svg";
@@ -9,17 +8,11 @@ import Sort from "@/assets/images/icons/ic_sort.svg";
 import { useEffect, useState } from "react";
 import styles from "@/styles/boards.module.css";
 import { Article } from "@/types/article";
+import BestArticleSection from "@/components/BestArticleSection";
 
 export default function CommunityFeedPage() {
-  const [bestArticles, setBestArticles] = useState<Article[]>([]);
-  const [allArticles, setAllArticles] = useState<Article[]>([]);
 
-  async function getBestArticle() {
-    const response = await axios.get("articles/?pageSize=3&orderBy=like");
-    const nextBestArticle = response.data.list ?? [];
-    setBestArticles(nextBestArticle);
-    console.log(nextBestArticle);
-  }
+  const [allArticles, setAllArticles] = useState<Article[]>([]);
 
   async function getAllArticle() {
     const response = await axios.get("articles");
@@ -29,42 +22,12 @@ export default function CommunityFeedPage() {
   }
 
   useEffect(() => {
-    getBestArticle();
     getAllArticle();
   }, []);
 
   return (
     <>
-      <div className={styles.articleContainer}>
-        <div className={styles.articleHeader}>
-          <h1 className={styles.articleTitle}>베스트 게시글</h1>
-        </div>
-        <div className={styles.bestArticleList}>
-          {bestArticles?.map((bestArticle) => (
-            <div key={bestArticle.id} className={styles.bestArticleWrapper}>
-              <Image
-                src={Badge}
-                alt="베스트게시글 뱃지"
-                style={{ position: "absolute", top: 0, left: 24 }}
-              />
-              <div className={styles.bestArticleContent}>
-                <div className={styles.articleTitle}>{bestArticle.title}</div>
-                <div className={styles.articleContentBottom}>
-                  <div className={styles.allArticleBottomLeft}>
-                    {bestArticle.writer.nickname}
-
-                    <div className={styles.likeCountWrapper}>
-                      <Image src={Heart} alt="좋아요 아이콘" />
-                      {bestArticle.likeCount}
-                    </div>
-                  </div>
-                  <div>{formatDate(bestArticle.createdAt)}</div>
-                </div>
-              </div>
-            </div>
-          ))}
-        </div>
-      </div>
+      <BestArticleSection/>
       <div className={styles.articleContainer}>
         <div className={styles.articleHeader}>
           <h1 className={styles.articleTitle}>게시글</h1>
